Fix warnings and inconsistencies across all source files
- Add missing headers: <fcntl.h>, <sys/stat.h>, <stdint.h> where needed
- Fix pointer format specifier: %x/(unsigned int)* -> %p/(void *) throughout
- Fix pthread function signatures to void *f(void *) as required by pthread_create
- Replace assert() on syscalls with proper perror()+exit(EXIT_FAILURE)
- Check stat() return value in getFilesize; make it static
- Fix int/void* size mismatch in thread-shared-vars via intptr_t
- Use EXIT_SUCCESS/EXIT_FAILURE consistently; add perror/usage messages
- Remove unused includes: <errno.h> in prodcons_mutex/sem, <string.h> in runner
- Remove redundant main() forward declarations in file_copy/file_copy_mmap
- Add pthread_mutex_destroy/cond_destroy on exit in goodcnt_mutex, prodcons_mutex
- Fix runner: pthread_exit(EXIT_SUCCESS) -> return NULL
- Fix shm: %i -> %d; fix sysV-msg: exit(-1) -> exit(EXIT_FAILURE)
- Remove stray semicolons after closing brace in ctshell and simple_pipe

+++ thread-shared-vars.c
@@ -1,4 +1,5 @@
 #include <pthread.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -7,7 +8,7 @@ char **ptr; /* global */
 
 /* thread routine */
 void *thread(void *vargp) {
-  int myid = (int)vargp;
+  int myid = (int)(intptr_t)vargp;
   static int cnt = 0;
 
   printf("[%d]: %s (cnt=%d)\n", myid, ptr[myid], ++cnt);
@@ -21,6 +22,6 @@ int main() { /* main thread */
   char *msgs[100] = {"Hello from foo", "Hello from bar"};
   ptr = msgs;
   for (i = 0; i < 2; i++)
-    pthread_create(&tid[i], NULL, thread, (void *)i);
+    pthread_create(&tid[i], NULL, thread, (void *)(intptr_t)i);
   pthread_exit(NULL);
 }
